Installed modules for typescript to work.
Honestly, it's a huge pain and I'm not sure everything is installed correctly,
might come back to it later.

Tried to structure the tests into folders, but mocha would just not
run. To run tests, follow the advice in readme and do `npm test
problem-number`.

Note for self: tsc will re-compile everything in the subdirectories, no
need to cd into them.